requirements.txt
fastapi
typer
bokeh
uvicorn
aiofiles
cartopy_offlinedata
datashader
graphene
# python-multipart
# xarray
# forest
# h5netcdf
# proj>=7.1.0
# pyproj>=2.6
# nodejs>=14.11.0
